Add emoji-based role management
Based off this discussion in #server-suggestions:
https://discord.com/channels/782201995011817493/793224619012128779/819308969460891648
Which concerns adding channels hidden behind roles and a method
through which to assign those roles

This feature will allow any moderator to create a post (possibly
in the welcome channel) which can be used to auto-assign roles
based on which reactions a user clicks

Example usage:
Making a post like this:
```
!role message
📖: @ReadingRole - For people who want to view the reading channel
✍🏾: @WritingRole - For people who want to view the writing channel
👂: @ListeningRole - For people who want to view the listening channel
```
Would result in the bot auto-adding those 📖, ✍🏾 and 👂 reactions to
the message.
Any time a user clicks one of the reactions they are assigned to
the corresponding role by the bot.
They may also remove their reaction to be unnassigned from that role

The usage rules currently:
- The message must start with "!role message"
- The reaction must be part of the standard set (no custom server
emojis)
- The sequence to form a role/reaction pair is:
{Emoji}{colon}{Role mention}
- Any line that contains the sequence is picked up and parsed
  - The line may include additional details after that sequence
  - Additional lines may be included that don't contain the
sequence (these are ignored)
  - The message can be edited after creation to have none of these
details and will still function as a role assignation message

Tables added to the database:
- RoleMessage
  - messageId
  - channelId
  - roleReactions (Array of objects:)
    - reactionName (emoji character code)
    - roleId (corresponding role for emoji)

Dev dependencies added to package.json:
- @babel/plugin-proposal-unicode-property-regex
  - Required to use Unicode property regexes
    - Allows us to use \p{Emoji} as a pattern to match any emoji

New events added to index.js:
- messageDelete
  - Fires whenever a message is deleted
    - Used to trigger the deletion of a role message from the database